Makefile corrections/split functions to files/check compatibility
